// old/src/lib/types.ts
export interface categoryInt {
  id: number;
  name: string;
  category_slug: string;
  slug: string;
  status?: boolean;
  total?: number;
  priority: number;
  deleted_at: null;
  created_at: string;
  updated_at: string;
}
// ...
export interface SearchParams {
  page?: string;
  slug?: string;
  category?: string;
  brand?: string;
  price?: string;
  searchterm?: string;
  paginate?: string;
  locale?: string;
}